+++ index.js
@@ -40,3 +40,4 @@ exports.constant = function(value) { return exports.identity.bind(null, value) }
 exports.clone = function(obj) { return obj ? exports.assign({}, obj)  : obj }
 exports.isBoolean = function(value) { return typeof value == "boolean" }
 exports.isNumber = function(value) { return typeof value == "number" }
+exports.isString = function(value) { return typeof value == "string" }